+++ src/renderer/App.tsx
@@ -14,6 +14,7 @@ import { useTranscription } from "./hooks/useTranscription";
 import { useSession } from "./hooks/useSession";
 import { useAudioPlayer } from "./hooks/useAudioPlayer";
 import { DownloadManagerDialog } from "./components/DownloadManagerDialog";
+import { ImportManagerDialog } from "./components/ImportManagerDialog";
 import { useAudioDownloads } from "./hooks/useAudioDownloads";
 import { useSettings } from "./hooks/useSettings";
 import { useModelDownloads } from "./hooks/useModelDownloads";
@@ -248,6 +249,13 @@ function App(): React.JSX.Element {
     handleRegenerateSubtitles,
     handleCancelRegeneration,
     handleUploadAudio,
+    handleDropAudioFiles,
+    importRecords,
+    isImporting,
+    showImportManager,
+    handleOpenImportManager,
+    handleCloseImportManager,
+    handleImportSelectSession,
   } = sessionMgmt;
 
   // Populate ref bridges for transcription/VAD callbacks
@@ -357,7 +365,7 @@ function App(): React.JSX.Element {
           onRegenerateSubtitles={handleRegenerateSubtitles}
           onCancelRegeneration={handleCancelRegeneration}
           onOpenFolder={(id) => window.capty.openAudioFolder(id)}
-          onUploadAudio={handleUploadAudio}
+          onUploadAudio={handleOpenImportManager}
           onDownloadAudio={() => setShowDownloadManager(true)}
           downloadBadge={downloadBadge}
           onAiRename={
@@ -593,6 +601,16 @@ function App(): React.JSX.Element {
           initialTab={settingsInitialTab}
           onTabChange={setSettingsInitialTab}
           onClose={() => setShowSettings(false)}
+        />
+      )}
+      {showImportManager && (
+        <ImportManagerDialog
+          records={importRecords}
+          isImporting={isImporting}
+          onUpload={handleUploadAudio}
+          onDropFiles={handleDropAudioFiles}
+          onSelectSession={handleImportSelectSession}
+          onClose={handleCloseImportManager}
         />
       )}
       {showDownloadManager && (